accept URL as well as domain name for --mirror argument
do not assume that the mirror mirrors the full directory structure of
nodejs.org, actually, some mirros just have the part of
`https://nodejs.org/download/release/` mirrored. for instance,
https://npm.taobao.org/mirrors/node, and do not assume that the mirror
always put the nodejs mirror in its `download/` directory.

in this change, a URL is accepted as well as a domain name. if a mirror
does not have "://" in it is specified, the default
"https://<domain_name>/download/releases" URL is used.

this change addresses the comment of
https://github.com/ekalinin/nodeenv/pull/208#issuecomment-380434680

also, ignore E127 for multi-line `with` statement, see discussion at
https://github.com/PyCQA/pycodestyle/issues/316, which is related to
E126, but it also applies this case.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>