support quotes & credentials in mirror config
this change performs two parts. The first is removing any extra
quotes from the configuration file so the examples in the README
work as expected.

The second is supporting HTTP Basic Authentication credentials within
the mirror argument itself so for folks fetching from a source that
requires authentication they are able to do so.