Use sh instead of bash
Currently bash is used to run some pretty trivial commands, but these
commands don't require anything bash-specific and can run with a regular
POSIX shell.

Use sh instead, to remove bash as a dependency for this package. There
is no need to update the list of dependencies since bash was already
missing from this list.